Fix "Play Again" text drift
Clicking outside of the "Play Again" button was causing the text to drift
down the screen.  The text is shifted up when the button is pressed and shifted
back when released.  On release, I wasn't checking if the button was pressed
in the first place before shifting the text back.