Make the samples work in portrait orientation
For most of the applications, it was a simple matter of setting manifest
properties and bar descriptor files.  For the VideoWindow and VideoPlayback
applications, we needed to play the video at the correct aspect ratio at
The largest size possible while still showing the whole area of the video.

Also, some of the GLES20 related code in VideoWindow application had to be
corrected.

Did not change HelloWorldDisplay since it is not suited for portrait
orientation.